Release v2.5
# Correlation Recorder Plugin v2.5 Release Note

We're excited to bring you the latest enhancements and features in the Correlation Recorder Plugin v2.5
Here's what's new:

## JSON Support on Automatic comparison and variable detection method

Not everything is a regex, that's why it's possible to record a JSON flow and let the automatic correlation recorder handle the dynamic elements.

In this release, there is a new addition, every time the plugin encounters a dynamic value in a JSON response (possible extractor) or JSON request (possible replacement) it will be translated to a [JSON Extractor](https://jmeter.apache.org/usermanual/component_reference.html#JSON_Extractor) with it's JSON Path auto generated.

## Other Improvements

- Fixed an issue that avoided consuming boundary length for automatic regex generation
- Fixed an issue that avoided saving a template when fresh start the plugin
- Fixed an issue that showed suggestions but weren't the ones applied
- Improvement in Github pages documentation