Release v3.0
# Auto Correlation Recorder Plugin v3.0 Release Note

We're excited to bring you the latest enhancements and features in the Auto Correlation Recorder Plugin v3.0
Here's what's new:

## Siebel removed from core

The following components were removed from the plugin
- Siebel Array JMeter Function
- Siebel Counter Correlation Replacement
- Siebel Row Correlation Extractor
- Siebel Row Id Correlation Replacement
- Siebel Row Params Correlation Replacement
The main reason this release is major one is due to the lack of backward compatibility.

However, not all are bad news. Now it's possible to access to a better Siebel correlation support by joining the [Blazemeter community](https://www.blazemeter.com/).
It's possible to use the new template and components directly from the plugin!

## Other Improvements

- Fix issue which reseted the JMeter templates.xml
- Allow to display templates description that contain custom rules despite of jar installation
- Dynamic class loader for custom extractors and replacements
- Custom extensions are now saved in lib/ext to benefit from JMeter class loading system
- Fix several issues related to apply suggestions based on a template